Complete Phase 2 of the refactor plan for the permission system
This commit finalizes the implementation of the permission system refactor, enhancing the architecture by moving the `IPermissionService` to the Infrastructure layer and ensuring all UI components correctly utilize the new permission checks. Key achievements include:

- Full implementation of the `PermissionService` with support for asynchronous permission checks and type-safe access rights.
- Updated dependency injection configuration to register the new permission service.
- Comprehensive testing of permission checks across various UI pages, ensuring compliance with Clean Architecture principles.

Additionally, the commit removes outdated access rights classes, streamlining the security model and improving maintainability. This aligns with the ongoing efforts to enhance code organization and adherence to architectural guidelines.
