+++ REFACTOR-PLAN.md
@@ -75,11 +75,27 @@ Application.Common.Interfaces.MultiTenant
   - [x] 创建IPermissionService接口
 - [x] 1.5 修复UI层DbContext直接继承问题
 
-### Phase 2: 权限系统重构 🔄 **待实现**
-- [ ] 2.1 在Infrastructure层实现 `IPermissionService`
-- [ ] 2.2 创建具体的权限服务实现
-- [ ] 2.3 更新依赖注入配置
-- [ ] 2.4 测试权限检查功能
+### Phase 2: 权限系统重构 ✅ **已完成**
+- [x] 2.1 在Infrastructure层实现 `IPermissionService`
+  - [x] PermissionService类完整实现
+  - [x] 支持HasPermissionAsync权限检查
+  - [x] 支持GetAccessRightsAsync类型安全访问
+  - [x] 基于反射的命名约定映射
+  - [x] 并发权限检查优化
+- [x] 2.2 创建具体的权限服务实现
+  - [x] 完整的权限检查逻辑
+  - [x] 用户权限获取功能
+  - [x] 系统所有权限枚举
+  - [x] 强类型AccessRights支持
+- [x] 2.3 更新依赖注入配置
+  - [x] services.AddScoped<IPermissionService, PermissionService>()
+  - [x] Infrastructure.DependencyInjection配置完成
+- [x] 2.4 测试权限检查功能
+  - [x] Products页面权限验证
+  - [x] Users页面权限验证  
+  - [x] Documents页面权限验证
+  - [x] Roles页面权限验证
+  - [x] 所有UI层正确通过接口调用
 
 ### Phase 3: 数据访问层隔离 ⏳ **待开始**
 - [ ] 3.1 确保所有数据访问都通过CQRS模式
@@ -171,7 +187,7 @@ using Infrastructure.PermissionSet;
 | 阶段 | 任务 | 状态 | 负责人 | 完成日期 |
 |------|------|------|--------|----------|
 | Phase 1 | Constants迁移 | ✅ 已完成 | AI Assistant | 2025-01-17 |
-| Phase 2 | 权限系统重构 | 🔄 进行中 | - | - |
+| Phase 2 | 权限系统重构 | ✅ 已完成 | AI Assistant | 2025-01-17 |
 | Phase 3 | 数据访问隔离 | ⏳ 待开始 | - | - |
 | Phase 4 | 服务接口化 | ⏳ 待开始 | - | - |
 | Phase 5 | 扩展方法优化 | ⏳ 待开始 | - | - |
@@ -215,6 +231,40 @@ Application/Common/
 │       └── AllAccessRights.cs
 └── Interfaces/
     └── IPermissionService.cs (新增)
+```
+
+## ✅ **Phase 2 重要成就**
+
+### 🎯 **完善的权限架构**
+1. **接口定义**: 在Application层创建了完整的IPermissionService接口
+2. **具体实现**: 在Infrastructure层实现了高性能的PermissionService
+3. **依赖注入**: 正确配置了服务注册，完全符合Clean Architecture
+4. **UI层集成**: 所有页面都正确通过接口使用权限服务
+
+### 🏗️ **技术特性实现**
+1. **反射机制**: 基于命名约定自动映射权限到AccessRights类
+2. **并发优化**: 权限检查使用并发任务，提高性能
+3. **类型安全**: 强类型的AccessRights避免了magic string
+4. **缓存友好**: 与现有的AuthenticationStateProvider和授权系统无缝集成
+
+### 📊 **架构合规验证**
+1. **完全合规**: 无任何UI → Infrastructure直接引用
+2. **依赖倒置**: UI层只依赖Application层的接口
+3. **单一职责**: 权限服务职责清晰，只处理权限相关逻辑
+4. **开闭原则**: 可轻松添加新的权限类型和AccessRights类
+
+### 💡 **实现亮点**
+```csharp
+// 🌟 类型安全的权限检查
+_accessRights = await PermissionService.GetAccessRightsAsync<ProductsAccessRights>();
+
+// 🌟 基于反射的自动映射
+// ProductsAccessRights.Create → Permissions.Products.Create
+
+// 🌟 并发权限检查优化
+var tasks = properties.ToDictionary(prop => prop, 
+    prop => _authService.AuthorizeAsync(user, $"Permissions.{sectionName}.{prop.Name}"));
+await Task.WhenAll(tasks.Values);
 ```
 
 ## 🔄 回滚计划
