--- src/Application/Common/Security/AccessRights/AllAccessRights.cs
@@ -1,69 +1,0 @@
-// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-
-namespace CleanArchitecture.Blazor.Application.Common.Security;
-
-public class PicklistSetsAccessRights
-{
-    public bool View { get; set; }
-    public bool Create { get; set; }
-    public bool Edit { get; set; }
-    public bool Delete { get; set; }
-    public bool Search { get; set; }
-    public bool Export { get; set; }
-    public bool Import { get; set; }
-}
-
-public class LogsAccessRights
-{
-    public bool View { get; set; }
-    public bool Search { get; set; }
-    public bool Export { get; set; }
-    public bool Purge { get; set; }
-}
-
-public class UsersAccessRights
-{
-    public bool View { get; set; }
-    public bool Create { get; set; }
-    public bool Edit { get; set; }
-    public bool Delete { get; set; }
-    public bool Search { get; set; }
-    public bool Import { get; set; }
-    public bool Export { get; set; }
-    public bool ManageRoles { get; set; }
-    public bool RestPassword { get; set; }
-    public bool SendRestPasswordMail { get; set; }
-    public bool ManagePermissions { get; set; }
-    public bool Deactivation { get; set; }
-    public bool ViewOnlineStatus { get; set; }
-}
-
-public class AuditTrailsAccessRights
-{
-    public bool View { get; set; }
-    public bool Search { get; set; }
-    public bool Export { get; set; }
-}
-
-public class LoginAuditsAccessRights
-{
-    public bool View { get; set; }
-    public bool Search { get; set; }
-    public bool Export { get; set; }
-    public bool ViewAll { get; set; }
-}
-
-public class TenantsAccessRights
-{
-    public bool View { get; set; }
-    public bool Create { get; set; }
-    public bool Edit { get; set; }
-    public bool Delete { get; set; }
-    public bool Search { get; set; }
-}
-
-public class DashboardsAccessRights
-{
-    public bool View { get; set; }
-}
